feat: organize views in different files
