+++ AuTime/View/Subviews/ActivitiesByCategory.swift
@@ -1,0 +1,20 @@
+//
+//  ActivitiesByCategory.swift
+//  AuTime
+//
+//  Created by Matheus Andrade on 19/11/21.
+//
+
+import SwiftUI
+
+struct ActivitiesByCategory: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct ActivitiesByCategory_Previews: PreviewProvider {
+    static var previews: some View {
+        ActivitiesByCategory()
+    }
+}
