+++ AuTime/View/Subviews/SearchBarView.swift
@@ -1,0 +1,41 @@
+//
+//  SearchBar.swift
+//  AuTime
+//
+//  Created by Matheus Andrade on 19/11/21.
+//
+
+import SwiftUI
+
+struct SearchBar: View {
+    @Binding var text: String
+    @State private var isEditing = false
+    
+    var body: some View {
+        HStack {
+            
+            TextField("Search ...", text: $text)
+                .padding(7)
+                .padding(.horizontal, 25)
+                .background(Color(.systemGray6))
+                .cornerRadius(8)
+                .padding(.horizontal, 10)
+                .onTapGesture {
+                    self.isEditing = true
+                }
+            
+            if isEditing {
+                Button(action: {
+                    self.isEditing = false
+                    self.text = ""
+                    UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+                }) {
+                    Text("Cancel")
+                }
+                .padding(.trailing, 10)
+                .transition(.move(edge: .trailing))
+                .animation(.default)
+            }
+        }
+    }
+}
